Add support for different media types in getRelease
Updated ReleaseController.kt to support either PDF or image/svg+xml as output formats for the getRelease method. `PDF` is the default when no `type` is specified. The change is aimed to enhance the application flexibility and widen the user options. Furthermore, a minor adjustment was made to templates/release/filled.ftlh to accurately display the url. Also, the necessary configurations were added for web resources and content compression in application.yaml.