Add Roadmap feature with scalable timeline
Integrated a new feature in the application that allows users to build a roadmap with a scalable timeline. The user interface for the feature has been implemented using a form in Freemarker and the endpoint has been set up in the MainController with performance metrics monitoring. Included the feature in the main navigation menu for easy access. The roadmap image generation has been updated to provide the capability to scale timelines according to user input. This feature was added to enhance the project planning functionalities of the application.