Update roadmaps and release strategy with improved visuals
This commit implements changes to the method of displaying texts in roadmaps and the animation of release strategy milestones. The 'wrapText' function was updated in 'RoadMapMaker.kt' to join all text fragments with an empty separator, making long roadmap texts more visible. Also, an animateMotion element was added in 'ReleaseRoadMapMaker.kt' to give a motion effect to release milestones based on their type, enhancing the visual representation of release strategies. Furthermore, changed color code in test.http for visual testing.