Remove unused code and refine implementations
Removed the "BadgeTest.kt" as it was not being used anymore, in the "RoadMapMaker.kt" class the "title" function was slighty refurbished to better handle SVG generation. Changes in "ReleaseTimelineMaker.kt" involve formating and encapsulation of fields, as well as an extraction of color logic for maintability. "ReleaseTimelineGroupedMaker.kt" was also modified for maintainability and to fit the project formating standards.