Update color scheme and refine visual elements in RoadMapMaker
The commit updates the color scheme of visual elements in the RoadMapMaker module by replacing previous color selections with new gradient fill values. Additionally, grid lines were enhanced for better visibility, and a redundancy in the ReleaseTimelineSummaryMaker was eliminated, improving the overall legibility of the roadmap. Lastly, the roadmap now provides functionality to copy Url for sharing purposes. This increases user convenience by allowing easy distribution of the roadmap view.