Add new APIs, update various UI components and add utility functions
Added new APIs for timeline, roadmap and release operations in oas.json. Updated the text wrapping feature and added new UI components like, labels, div tags, and buttons in tm.ftlh. These changes are made to enhance the user interface and to better illustrate the timelines and roadmaps. Furthermore, utility functions to convert specific texts to spans, to handle the display configuration and to build a road have been added in TimelineMaker.kt and Entry.kt, improving the functionality and usability.