Implement rectangle button shape and styles for DocOps Extension support
This commit adds functionality to render rectangle button shapes for the DocOps Extension support by implementing required classes and functions. Firstly, a new Rectangle.kt file was introduced. This class extends the AbstractButtonShape and provides unique methods for a rectangular shape. Secondly, this commit also includes predefined styles in PredefinedStyle.kt such as linkText, myBox and keyFrame necessary for SVG presentation of rectangle button. This enhancement was done to expand the versatility of available button shapes in the extension, particularly for users who specifically require rectangular shapes for their use-case.