Refactor button handling and introduce PNG export functionality
This commit revolves around refactoring the code for creating buttons and introducing a new functionality to export them as PNGs. Previously, button creation code was rigid and was only capable of SVG exports. Now, the create() function has been modified to make use of a new createSVGShape() function, which returns a ButtonShape that can be drawn in different formats based on input.

We've also introduced changes across different button shapes (Regular, Round, and Large), distinguishing between PDF and non-PDF behaviors. This was done to ensure cleaner display of buttons without redundancies like filters and gradients while exporting in PDF format.

To support PNG exports, new end-points "/buttons/png" and "/buttons/form/png" are added to the ButtonController. We've also added a new template 'formjsontopng.ftlh' to aid this functionality.

This refactor and introduction of new functionality improve the flexibility and usability of button generation. It further enhances the offering of our service, allowing users to choose between different formats as per requirement.