Add font color field to FormBadge
Added a new field named 'fontColor' to FormBadge class to allow customization of font color in the Badge. Updated the FormBadge class, BadgeController, DocOpsBadgeGenerator, and badge.ftlh template to reflect this change. This allows users to now input their choice of font color while creating badges and the color is applied to the text on the badge.