Optimize ScorecardController methods and add log info
Refactored the methods in the ScorecardController class to measure the execution time and log this information, so as to monitor performance. Methods include getScoreCard, putScorecard, and fromJsonToScorecard. Also, added a new scorecard sample 'score2.json'. Updated index.html to include links to scorecard samples in both dark and light themes. Altered ScoreCardMaker class and MainController to accommodate these changes. It improves the tracking and understanding of method executions for better performance tuning.