Update timing log and dependency versions
Introduced timing logs in different GET and PUT methods in `TimelineController.kt`, `ReleaseController.kt`, and `RoadmapPlanController.kt`, to track and improve performance. Upgraded HTML manipulation library 'htmx' from 1.9.2 to 1.9.6 and JQuery from 2.1.1 to 3.7.1 in `index.html`. Further, updated the Dokka, and JVM versions in `pom.xml` for keeping up-to-date with the latest improvements and updates.