Add pink theme and extend color palette
A new scorecard theme 'Pink' was added and the color palette in clrs.js was extended. This provides more variety for users when setting up their scorecards. The new theme was added to index.html for user selection. Meanwhile, the color palette specified in 'baby-boy.json' and 'score-green.json' was also updated to reflect the additional colors. This aesthetic update is aimed to enhance user experience with richer choices for customization.