Update roadmap builder and improve template layout
This commit introduces a change to the roadmap builder which now allows for greater control over gradient indexing. The template and visual layout of forms have also been improved to enhance clarity and user interaction. Adjustments were made to the aesthetic aspects of the Boxy, Timeline and other templates for better visibility and comprehension.