Update button design and positioning
The styling of buttons has been upgraded for better visual design, the padding for button was increased from 5 to 10 to accommodate the new design. Also enhanced Regular and Round button classes with new button generation logic; Round class has additional logic to handle button rows and introduces ItemCount class for calculating the button index. This improves scalability and visual appeal of buttons on the platform.