Update ScoreCard handling and improve project dependencies
The significant changes in this commit revamp the ScoreCard handling in the Kotlin code. The ScoreCard now calculates its height based on initiative and outcome items. We also allow the ScoreCard items to convert their display text into a list based on a given size. In terms of project dependencies, we have updated the Kotlin version and improved AspectJ configuration.