Update ScoreCardMaker styles and elements
The ScoreCardMaker.kt has been updated with modifications in styling and elements used. Certain styles and elements were removed to simplify the design, while notable enhancements included the introduction of new linear gradients and adjustments to the positioning and colouring of elements. Additional changes also encompass the replacement of the previous background with a rectangle filled with "#F7F7F7".